production ready build features file and README update
